refactor(arduino): split peripherals into headers
